thiếu login, auth, secu

+++ lms_be/backend_lms/src/main/java/com/example/backend_lms/service/ConversationService.java
@@ -4,7 +4,9 @@ import com.example.backend_lms.dto.ConversationDTO;
 import com.example.backend_lms.dto.PageDTO;
 import com.example.backend_lms.entity.Conversation;
 import com.example.backend_lms.repo.ConversationRepo;
+import com.example.backend_lms.repo.UserRepo;
 import jakarta.transaction.Transactional;
+import javassist.NotFoundException;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -13,6 +15,7 @@ import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
@@ -20,6 +23,9 @@ public class ConversationService {
 
     @Autowired
     ConversationRepo conversationRepo;
+
+    @Autowired
+    UserRepo userRepo;
 
     public ConversationDTO convert(Conversation conversation){
         return new ModelMapper().map(conversation, ConversationDTO.class);
@@ -31,8 +37,15 @@ public class ConversationService {
     }
 
     @Transactional
-    public void delete(int id){
-        conversationRepo.deleteById(id);
+    public void delete(int id, int user_id) throws NotFoundException {
+        if(userRepo.findById(user_id).isPresent() && conversationRepo.findById(id).isPresent()){
+            if(Objects.requireNonNull(conversationRepo.findById(id).orElse(null)).getUser().getId() == user_id){
+                conversationRepo.deleteById(id);
+            }
+        }
+        else{
+            throw new NotFoundException("Id người dùng hoặc bình luận không hợp lệ");
+        }
     }
 
     public PageDTO<List<ConversationDTO>> getConversation(int course_id, int current_page){
